Remove unnecessary test file, move makestyles back to component files, keep App.js clean and move moment call outside of useEffect
