+++ client/src/components/ToDoList/AddToDo/index.js
@@ -1,9 +1,38 @@
 import React, { useState } from "react";
 import PropTypes from "prop-types";
 import moment from "moment";
-import { Button, CardActions, CardHeader, TextField } from "@material-ui/core";
+import { Button, CardActions, CardHeader, TextField, makeStyles } from "@material-ui/core";
 import styles from "./style.module.scss";
-import useStyles from "./styles";
+
+const useStyles = makeStyles({
+    todoListTitle: {
+        textAlign: "center",
+        backgroundColor: "#30A0F5",
+        color: "white",
+        marginBottom: "1.5vh",
+        paddingTop: "2.3vh",
+        paddingBottom: "2.3vh"
+    },
+    todoInputTextField: {
+        [`& fieldset`]: {
+            borderRadius: "40px"
+        }
+    },
+    primaryButton: {
+        backgroundColor: "#30A0F5",
+        color: "white",
+        borderRadius: "20px" 
+    },
+    cardActions: {
+        justifyContent: "flex-end"
+    },
+    cancelButton: {
+        backgroundColor: "white",
+        color: "black",
+        borderColor: "#30A0F5",
+        borderRadius: "20px"
+    }
+});
 
 function AddToDo({ cancelClicked, addClicked }) {
     const currentDate = moment().format("YYYY-MM-D");
